Make directory.HandleOps() unexported method of keyserver/listener.go - HandleOps is really a server-level function. This change makes the directory API strictly about operations on the data structure.
