+++ protocol/directory_test.go
@@ -285,58 +285,6 @@ func TestDirectoryKeyLookupInEpoch(t *testing.T) {
 	}
 }
 
-func TestDirectoryKeyLookupInEpochBadEpoch(t *testing.T) {
-	N := 3
-
-	d, _ := NewTestDirectory(t, false)
-	for i := 0; i < N; i++ {
-		d.Update()
-	}
-
-	// Send an invalid KeyLookupInEpochRequest (epoch > d.LatestEpoch())
-	// Expect ErrorMalformedClientMessage
-	req := &Request{
-		Type:    KeyLookupInEpochType,
-		Request: &KeyLookupInEpochRequest{"alice", uint64(6)},
-	}
-
-	_, err := d.HandleOps(req)
-	if err != ErrorMalformedClientMessage {
-		t.Fatal("Expect error", ErrorMalformedClientMessage, "got", err)
-	}
-}
-
-func TestMonitoringBadStartEpoch(t *testing.T) {
-	N := 3
-
-	d, _ := NewTestDirectory(t, false)
-	for i := 0; i < N; i++ {
-		d.Update()
-	}
-
-	// Send an invalid MonitoringRequest (startEpoch > d.LatestEpoch())
-	// Expect ErrorMalformedClientMessage
-	req := &Request{
-		Type:    MonitoringType,
-		Request: &MonitoringRequest{"alice", uint64(6), uint64(10)},
-	}
-	_, err := d.HandleOps(req)
-	if err != ErrorMalformedClientMessage {
-		t.Fatal("Expect error", ErrorMalformedClientMessage, "got", err)
-	}
-
-	// Send an invalid MonitoringRequest (startEpoch > EndEpoch)
-	// Expect ErrorMalformedClientMessage
-	req = &Request{
-		Type:    MonitoringType,
-		Request: &MonitoringRequest{"alice", uint64(2), uint64(0)},
-	}
-	_, err = d.HandleOps(req)
-	if err != ErrorMalformedClientMessage {
-		t.Fatal("Expect error", ErrorMalformedClientMessage, "got", err)
-	}
-}
-
 func TestPoliciesChanges(t *testing.T) {
 	d, _ := NewTestDirectory(t, false)
 	if p := d.LatestSTR().Policies.EpochDeadline; p != 1 {
